Remove read_private_key
