+++ examples/demo-stf/src/tests/stf_tests.rs
@@ -1,17 +1,17 @@
 #[cfg(test)]
 pub mod test {
 
+    use sov_cli::wallet_state::PrivateKeyAndAddress;
     use sov_data_generators::bank_data::get_default_token_address;
     use sov_data_generators::{has_tx_events, new_test_blob_from_batch};
     use sov_modules_api::default_context::DefaultContext;
     use sov_modules_api::default_signature::private_key::DefaultPrivateKey;
-    use sov_modules_api::{PrivateKey, WorkingSet};
+    use sov_modules_api::{Context, PrivateKey, WorkingSet};
     use sov_modules_stf_template::{Batch, SequencerOutcome};
     use sov_rollup_interface::mocks::{MockBlock, MockDaSpec, MOCK_SEQUENCER_DA_ADDRESS};
     use sov_rollup_interface::stf::StateTransitionFunction;
     use sov_state::ProverStorage;
 
-    use crate::genesis_config::read_private_key;
     use crate::runtime::Runtime;
     use crate::tests::da_simulation::simulate_da;
     use crate::tests::{create_new_app_template_for_tests, get_genesis_config_for_tests, C};
@@ -216,5 +216,26 @@ pub mod test {
 
         // Assert that there are no events
         assert!(!has_tx_events(&apply_blob_outcome));
+    }
+
+    pub fn read_private_key<C: Context>() -> PrivateKeyAndAddress<C> {
+        let token_deployer_data =
+            std::fs::read_to_string("../test-data/keys/token_deployer_private_key.json")
+                .expect("Unable to read file to string");
+
+        let token_deployer: PrivateKeyAndAddress<C> = serde_json::from_str(&token_deployer_data)
+            .unwrap_or_else(|_| {
+                panic!(
+                    "Unable to convert data {} to PrivateKeyAndAddress",
+                    &token_deployer_data
+                )
+            });
+
+        assert!(
+            token_deployer.is_matching_to_default(),
+            "Inconsistent key data"
+        );
+
+        token_deployer
     }
 }
